Extract generic HistoryView from snippets history view
