+++ wagtail/admin/templates/wagtailadmin/generic/history_results.html
@@ -1,0 +1,6 @@
+{% extends "wagtailadmin/generic/index_results.html" %}
+{% load i18n %}
+
+{% block no_results_message %}
+    <p>{% trans "No log entries found." %}</p>
+{% endblock %}
