+++ wagtail/admin/templates/wagtailadmin/shared/header.html
@@ -75,7 +75,7 @@
                     <div class="actionbutton">
                         {{ base_actions }}
                         {% with action_icon|default:'plus' as action_icon_name %}
-                            <a href="{{ action_url }}" class="button bicolor button--icon">{% icon name=action_icon_name wrapped=1 %}{{ action_text }}</a>
+                            <a href="{{ action_url }}" class="button bicolor button--icon">{% icon name=action_icon_name wrapped=1 %}{{ action_text|capfirst }}</a>
                         {% endwith %}
                     </div>
                 {% endif %}
