+++ wagtail/snippets/templates/wagtailsnippets/snippets/history.html
@@ -1,23 +1,1 @@
-{% extends 'wagtailadmin/reports/base_report.html' %}
-{% load i18n wagtailadmin_tags %}
-
-{% block titletag %}{% blocktrans trimmed with title=subtitle %}Snippet history for {{ subtitle }}{% endblocktrans %}{% endblock %}
-{% block bodyclass %}model-{{ model_opts.model_name }}{% endblock %}
-
-{% block actions %}
-    <a href="{% url view.edit_url_name object.pk|admin_urlquote %}" class="button bicolor button--icon">{% icon name="edit" wrapped=1 %}{% trans "Edit this snippet" %}</a>
-{% endblock %}
-
-{% block content %}
-    {% include 'wagtailadmin/shared/headers/slim_header.html' %}
-
-    {{ block.super }}
-{% endblock %}
-
-{% block results %}
-    {% if object_list %}
-        {% component table %}
-    {% else %}
-        <p>{% trans "No log entries found." %}</p>
-    {% endif %}
-{% endblock %}
+{% extends 'wagtailadmin/generic/index.html' %}
